requirements.txt
nuscenes-devkit
numpy
matplotlib
pillow
open3d

